Migrate custom widget implementations
